Fix setter.rs tests and refactor instance tests
- Replace undefined create_constraint function with Constraint::equal_to_zero
- Update decision_variables and constraints to use btreemap\! macro
- Simplify objective function creation using linear\! macro
- Remove unused imports and clean up test code
- Add constructor methods to Constraint struct
